// src/redux/posts/types.ts
export enum postsTypes {
  FETCH_GET_POSTS = 'FETCH_GET_POSTS',
  FETCH_GET_POSTS_REQUEST = 'FETCH_GET_POSTS_REQUEST',
  FETCH_GET_POSTS_SUCCESS = 'FETCH_GET_POSTS_SUCCESS',
  FETCH_GET_POSTS_FAILURE = 'FETCH_GET_POSTS_FAILURE',

  FETCH_CREATE_POST = 'FETCH_CREATE_POST',
  FETCH_CREATE_POST_REQUEST = 'FETCH_CREATE_POST_REQUEST',
  FETCH_CREATE_POST_SUCCESS = 'FETCH_CREATE_POST_SUCCESS',
  FETCH_CREATE_POST_FAILURE = 'FETCH_CREATE_POST_FAILURE',

  FETCH_GET_POST_DETAIL = 'FETCH_GET_POST_DETAIL',
  FETCH_GET_POST_DETAIL_SUCCESS = 'FETCH_GET_POST_DETAIL_SUCCESS',
  FETCH_GET_POST_DETAIL_FAILURE = 'FETCH_GET_POST_DETAIL_FAILURE',
}